+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 _01.07.2026_
 
 - Fixed an issue, where the User was not able to acknowledge the sleep timer safety stop.
+- Sleep timer doesnt run during auto tune
 
 # Gluetex bugfixes
 _25.06.2026_